NDRS-833: Extract `LinearChain` state to a file.
Leaving only the component-level logic in `linear_chain.rs`.
